test/crimson: verify the futurization in an errorated future.
`seastar::future` allows a lambda passed to e.g. `then()` to
return not only a future but also any non-future type like
a plain integer:

  ```cpp
  seastar::now().then([] {
    return 42;
  }).then([] (int futurized_life) {
    // ...
  });
  ```

In such case, the plain type is being wrapped in a future.
This process is called _futurization_ and is performed by
`seastar::futurize` called from the internals of `future`.

In this commit we want to ensure analogous facility is offered
by errorated futures.

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>