doc/cephadm: enrich "service status"
This PR improves the syntax of the "Service
Status" section of the "Service Managment"
section of the cephadm guide. This includes
pretty significant reworking of the information
in the section, so vetting this one might be
annoying. Anyway, I think I've lowered the
cognitive load on the reader.

Signed-off-by: Zac Dover <zac.dover@gmail.com>
(cherry picked from commit 31558e5568da5b445b2da1b98ac2eb626d8efdf3)