cephadm: make src/cephadm/cephadm importable as module
Enables type checking for src/cephadm/cephadm tests.

Type checking for tests of src/cephadm/cephadm has basically been
disabled by the first line of src/cephadm/tests/test_cephadm, saying `#
type: ignore`.  All other type checks and resulting suggestions by the
IDE and testing tools related to types were basically disabled due to
that line.

Disabling type checking that way was at least convenient if not
necessary, as the types couldn't be provided because the file `cephadm`
couldn't be imported as a module.  A simple symlink from `cephadm` to
`cephadm.py` could have resolved the issue as well, but I considered
this to be a workaround.

The intention of having the file named `cephadm` seems to be that it is
being installed and wanted to be directly callabe.  This can also be
achieved by changing the installation instructions of the packages to be
created accordingly.  This is the approach taken here.

Signed-off-by: Patrick Seidensal <pseidensal@suse.com>