doc/dev: s/reposotory/repository/ (really)
This corrects the heinous misspelling described in the
substitution expression in the title. This misspelling is
all the more egregious because it appears in a title, and
therefore would be used to create links if it had not been
caught.

Signed-off-by: Zac Dover <zac.dover@gmail.com>
(cherry picked from commit 298b446c35d19ce43ede513a802d0655bcbdf82f)