nuke: kill -9 the teuthology process
If the process has been kill -STOPped, then we'll unlock the machines, but
the process will stick around and we'll try to nuke it again later,
zapping the machines after they're being used by some other job, leading
to failures.  (Usually this manifests as an error when the other job stops
where it has trouble gzipping the logs.)

Use -9 to make sure even STOPped processes are killed.

Signed-off-by: Sage Weil <sage@newdream.net>