include/cpp-btree: drop node handle support
this is a leftover in 84a91b71141fed9695c3ddb03568cd1c9041d9ac,

the node support is useful if we want to move a node around in a
node-based container. but in the case of b-tree, multiple elements
are packed into a single node/block, so we cannot benefit from this
feature, i think. so just drop all of them.

Signed-off-by: Kefu Chai <kchai@redhat.com>