crimson/osd: fix premature retirement of client req while sending OSDMap.
We need to go through the entire pipeline even when early exiting.
Otherwise we'll brake the ordering and blow up `OpSequencer`.

Similiar problem affects in the `is_pg_op() == true` branch which
will be fixed in a follow-up commit if this one works as expected.

Likely the EIO handling must be revised as well.

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>