mon/LogMonitor: externally log events from when we were out of quorum
Previously, we would externally log any item from the LogSummary.version
(written every commit, with version-1) and last_committed.  That meant
that we logged everything only if the mon was in quorum.  If we fell
out of quorum and then later caught up, we would/could miss items.

Instead, explicitly track what version we have externally logged to, and
log everything up until that point.  Only consule summary.version if we
don't have a local marker recorded (e.g., because we just upgraded).

Signed-off-by: Sage Weil <sage@newdream.net>