crush: eliminate min_size and max_size
We don't use min/max_size for anything any more.  Switch to encoding it
universally as 1 and 100, and ignoring when we decode.  We keep it around
for backward compat when (re)encoding without the SERVER_QUINCY feature.

Drop the crush_rule_mask struct since the only variable piece now is the
type.

Signed-off-by: Sage Weil <sage@newdream.net>