mgr/telemetry: added a new channel for perf counters
This commmit adds a new channel to the telemetry module called 'perf' that sums up all of the perf counter metrics across a cluster. Here is a rundown of the changes I made:
	- added the name of the channel to ALL_CHANNELS
	- added the perf channel under MODULE_OPTIONS (by default, the channel is on)
	- initialized the channel to True under __init__
	- wrote a new method called gather_perf_counters() that sums the metrics among all of the cluster's daemons and gets the info ready for the JSON report
	- added the channel to get_active_channels()
	- added perf counters to the JSON report by calling gather_perf_counters() under compile_report()

Signed-off-by: Laura Flores <lflores@redhat.com>
(cherry picked from commit b5005da5b01dc6a501da169747555800391718ca)

Conflicts:
	src/pybind/mgr/telemetry/module.py

Type-checking was not in pacific.