mgr/telemetry: turn perf channel off by default and decrement revision number
Originally, the perf channel was turned on by default, and the revision number was incremented. However, it is best for now that the perf channel is turned off by default and the revision number is returned to its previous state. This is to avoid having the telemetry module turn off once this PR is merged, which would likely result in losing many of our current telemetry users.

The perf channel will be updated to send more information in the future, so once it is finalized with the information we want to collect, we may reconsider turning the channel on and incrementing the revision number.

Signed-off-by: Laura Flores <lflores@redhat.com>
(cherry picked from commit 04ff1f8fb8dcbe1de28f4d4552c391aee0812126)

 Conflicts:
	src/pybind/mgr/telemetry/module.py

 There was no type-checking in pacific.