mgr/rook: create and use DefaultFetcher and LSOFetcher classes
This commit creates the DefaultFetcher and LSOFetcher classes to
fetch PVs and return a list of devices for gatherin inventory.

The LSOFetcher inherits from the DefaultFetcher class and is used if
the storage class specified in the config was created by the LSO.

The RookCluster uses these classes to gather inventory and pass a list
of Device objects to back to the RookOrchestrator.

Signed-off-by: Joseph Sawaya <jsawaya@redhat.com>