common: optimize objects memory allocations using pools
A significant time is spent in allocating memory for class objects on heap.
i.e., in calls to c++ operator new(long) and operator delete(void*).
This patch creates by-size pools to be used for allocating common objects,
which are relatively large, such as MOSDOp, Op, and librbd disaptch specs.

Signed-off-by: Or Ozeri <oro@il.ibm.com>